Drop separate PreProtocol object in favour of AgendaItem.
