+++ opengever/meeting/browser/meetings/templates/protocol.pt
@@ -14,10 +14,10 @@
 
             <metal:block fill-slot="actions">
 
-              <metal:block tal:repeat="protocol view/get_pre_protocols">
+              <metal:block tal:repeat="protocol view/get_agenda_items">
               <metal:block tal:define="name protocol/name">
 
-                <h2 class="protocol_title" tal:content="protocol/title"></h2>
+                <h2 class="protocol_title" tal:content="protocol/get_title"></h2>
 
                 <metal:block tal:condition="protocol/has_proposal">
                     <label tal:attributes="for string:${name}.legal_basis:record"
