+++ opengever/meeting/protocol.py
@@ -5,115 +5,9 @@ from zope.i18n import translate
 import json
 
 
-class PreProtocol(object):
-    """Abstract from Proposal and AgendaItem for protocol-related attributes.
-
-    Currently some attributes are available on a proposal wheras others are
-    stored directly on an agenda_item.
-
-    """
-    def __init__(self, agenda_item):
-        self._agenda_item = agenda_item
-        self._proposal = self._agenda_item.proposal
-
-    @property
-    def has_proposal(self):
-        return self._proposal is not None
-
-    @property
-    def legal_basis(self):
-        return self._proposal.legal_basis if self.has_proposal else None
-
-    @property
-    def initial_position(self):
-        return self._proposal.initial_position if self.has_proposal else None
-
-    @property
-    def considerations(self):
-        return self._proposal.considerations if self.has_proposal else None
-
-    @property
-    def proposed_action(self):
-        return self._proposal.proposed_action if self.has_proposal else None
-
-    @property
-    def discussion(self):
-        return self._agenda_item.discussion
-
-    @property
-    def decision(self):
-        return self._agenda_item.decision
-
-    @property
-    def name(self):
-        return "preprotocols.{}".format(self._agenda_item.agenda_item_id)
-
-    @property
-    def title(self):
-        return u"{} {}".format(self.number, self.description)
-
-    @property
-    def description(self):
-        return self._agenda_item.get_title()
-
-    @property
-    def number(self):
-        return self._agenda_item.number
-
-    def update(self, request):
-        """Update with changed data."""
-
-        data = request.get(self.name)
-        if not data:
-            return
-
-        if self.has_proposal:
-            self._proposal.legal_basis = data.get('legal_basis')
-            self._proposal.initial_position = data.get('initial_position')
-            self._proposal.considerations = data.get('considerations')
-            self._proposal.proposed_action = data.get('proposed_action')
-
-        self._agenda_item.discussion = data.get('discussion')
-        self._agenda_item.decision = data.get('decision')
-
-    def get_field_data(self, include_initial_position=True,
-                       include_legal_basis=True, include_considerations=True,
-                       include_proposed_action=True, include_discussion=True,
-                       include_decision=True):
-        data = {
-            'number': self.number,
-            'description': self.description,
-            'title': self.title,
-        }
-        if include_initial_position:
-            data['markdown:initial_position'] = self._sanitize_text(
-                self.initial_position)
-        if include_legal_basis:
-            data['markdown:legal_basis'] = self._sanitize_text(
-                self.legal_basis)
-        if include_considerations:
-            data['markdown:considerations'] = self._sanitize_text(
-                self.considerations)
-        if include_proposed_action:
-            data['markdown:proposed_action'] = self._sanitize_text(
-                self.proposed_action)
-        if include_discussion:
-            data['markdown:discussion'] = self._sanitize_text(self.discussion)
-        if include_decision:
-            data['markdown:decision'] = self._sanitize_text(self.decision)
-
-        return data
-
-    def _sanitize_text(self, text):
-        if not text:
-            return None
-
-        return text
-
-
 class PreProtocolData(object):
 
-    def __init__(self, meeting, pre_protocols=None,
+    def __init__(self, meeting, agenda_items=None,
                  include_initial_position=True, include_legal_basis=True,
                  include_considerations=True, include_proposed_action=True,
                  include_discussion=True, include_decision=True):
@@ -126,14 +20,7 @@ class PreProtocolData(object):
         self.include_decision = include_decision
 
         self.meeting = meeting
-        if pre_protocols:
-            self.pre_protocols = pre_protocols
-        else:
-            self.pre_protocols = []
-            for agenda_item in self.meeting.agenda_items:
-                if agenda_item.is_paragraph:
-                    continue
-                self.pre_protocols.append(PreProtocol(agenda_item))
+        self.agenda_items = agenda_items or self.meeting.agenda_items
 
         self.data = {}
         self.add_base()
@@ -183,9 +70,9 @@ class PreProtocolData(object):
 
     def add_agenda_items(self):
         self.data['agenda_items'] = []
-        for pre_protocol in self.pre_protocols:
+        for agenda_item in self.agenda_items:
             self.data['agenda_items'].append(
-                pre_protocol.get_field_data(
+                agenda_item.get_field_data(
                     include_initial_position=self.include_initial_position,
                     include_legal_basis=self.include_legal_basis,
                     include_considerations=self.include_considerations,
@@ -242,7 +129,7 @@ class ExcerptProtocolData(object):
         self.data['agenda_items'] = []
         for agenda_item in self.agenda_items:
             self.data['agenda_items'].append(
-                PreProtocol(agenda_item).get_field_data())
+                agenda_item.get_field_data())
 
     def as_json(self):
         return json.dumps(self.data)
